refactor: extract shared SineSynthVoice for realtime and offline paths
- Move SimpleSineVoice/SimpleSineSound out of AudioEngine into
  source/Audio/SineSynthVoice.h (header-only, JUCE SynthesiserVoice)
- Replace the duplicated OfflineSineVoice in WavFileExporter with the
  shared class so live and WAV-export paths use identical voices
- No behaviour change; verified via wsl-build, 51 test classes green,
  format check, clang-tidy (0 diagnostics) and win-build
